SwiGLU: move path selection to swiglu_op; swiglu_fw full-fusion only
- Path switch and SwigluFwPath enum live in swiglu_op (ops layer).
- swiglu_fw.cpp implements only full-fusion (calls ttml_swiglu_fw).
- GateUp = production default. Composite and FullFusion kept for tests/development.
- FullFusion: document saves memory, runtime suboptimal; log_warning in switch.
- Composite: log_warning in switch; comments that both are for further tests/dev.

Made-with: Cursor
